Spring Boot JPA Join Mapping

+++ Spring Data/Spring-Boot-Mappings/Spring-Boot-Mapping-ManyToMany/src/main/java/com/mylearning/springbootmappingmanytmany/entity/Course.java
@@ -1,5 +1,8 @@
 package com.mylearning.springbootmappingmanytmany.entity;
 
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -19,13 +22,13 @@ import java.util.List;
 public class Course {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String courseName;
     private String teacherName;
     private int modules;
     private double fees;
 
     @ManyToMany(mappedBy = "courses") // Inverse side, referring to the courses field in Student
+    // @JsonManagedReference // tum karo response me manage yeh hum isko bol rahe hai
     private List<Student> students = new ArrayList<>();
-
 }
